СОздание классов для отдельных объектов
