+++ .idea/Galaxy-War.iml
@@ -4,7 +4,7 @@
     <content url="file://$MODULE_DIR$">
       <excludeFolder url="file://$MODULE_DIR$/venv" />
     </content>
-    <orderEntry type="jdk" jdkName="Python 3.9 (Galaxy-War)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Python 3.8 (Galaxy-War)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
 </module>
